Update the build rule for pkg-config
For the purpose of make it works better in macOS, I wish to submit this
patch.

1. symlinks

In libmongocxx, the structure of the headers are seems as follow:

```
.
├── bsoncxx
│   └── v_noabi
│       └── bsoncxx
└── mongocxx
    └── v_noabi
        └── mongocxx
```

And the cflags are like this:

```
$ pkg-config --cflags libmongocxx
-I/usr/local/Cellar/mongo-cxx-driver/3.1.3/include/mongocxx/v_noabi \
  -I/usr/local/Cellar/mongo-c-driver/1.8.2/include/libmongoc-1.0 \
  -I/usr/local/Cellar/mongo-cxx-driver/3.1.3/include/bsoncxx/v_noabi \
  -I/usr/local/Cellar/mongo-c-driver/1.8.2/include/libbson-1.0
```

There are 2 header directory using basename `v_noabi` but contains
different subdirectories.

This patch will check the basename, and add a suffix for duplicated
basenames.

Besides, it will change nothing if there is no conflict basenames.

2. linkopts

If we try to parse the static libs in macOS, we may get result as
follow:

```
$ pkg-config --static --libs libmongocxx
-L/usr/local/Cellar/mongo-cxx-driver/3.1.3/lib \
  -L/usr/local/Cellar/mongo-c-driver/1.8.2/lib \
  -L/usr/local/Cellar/mongo-cxx-driver/3.1.3/lib \
  -L/usr/local/Cellar/mongo-c-driver/1.8.2/lib \
  -lmongocxx -lmongoc-1.0 -lsasl2 \
  -framework Security \
  -framework CoreFoundation \
  -lz -lbsoncxx -lbson-1.0
```

Note: the option '-framework' and 'Security/CoreFoundation' should be
connected in macOS.