add more xrefs in object literals
Summary:
An object literal, of the form
  let foo = { bar: 3 };
should define a "foo.bar" variable.

At the AST level, we parse this as
(1) a variable declaration of foo, which
(2) contains an anonymous object literal

So the full name ends up as something like "foo.anon0.bar".
But this is fine as long as the references to foo.bar also use
that name.

In fixing this, it revealed a bug in the function.ts test --
we were conflating the function param named 'num' with the property
named 'num' on the anonymous object within it.  Mark the problem
with a comment for now, as the new behavior is more correct.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2120