Make sed usage less GNU-specific.
Summary:
In the land of macOS, the ancient and venerable sed installed therein, served
as it has for aeons beyond count in valiant and faithful measure, doth not
concur with the GNU in the matter of command separators.

Forsooth! She cries: There shall be newlines.

And so there shall.

Reviewers: salguarnieri, #core_team

Reviewed By: salguarnieri, #core_team

Subscribers: #core_team

Maniphest Tasks: T285

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2123