Fix the protobuf dependency on zlib.
Summary:
When I updated the protobuf rules in D2122, I also inadvertently introduced a
dependency bug. In particular, I did not plumb through the correct zlib deps,
so that protobuf built using the ambient system-installed zlib instead of the
one we build in Bazel.

This change binds our //third_party/zlib rule as an external dependency and
fixes up the BUILD file for protobuf to use it properly.

Addresses T287.

Reviewers: zarko, #core_team

Reviewed By: zarko, #core_team

Subscribers: #core_team

Maniphest Tasks: T287

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2137