Update the output directories for pre-built Go protos.
Summary:
Now that we are using the upstream Go protobuf rules, the outputs are named in
harmony with the aspect, e.g., foo_go_proto instead of foo_proto.  Update the
tool to copy the generated results to a directory that matches the new package
name, to avert confusion in tools that want the names to match up.

Relocate the pre-generated Go protobuf packages.

To verify that the change in the updater script was correct, generate new proto
packages and remove the old ones. This commit collapses the two steps.

Reviewers: #core_team, benjyw

Reviewed By: benjyw

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2163