entryset: Split long symbols into smaller chunks.
Summary:
An entryset.Set can now take an optional maximum size for symbols, and symbol
values longer than that will be split so that after prefix coding no suffix
will be longer than the specified maximum. If no maximum size is set, or if the
maximum is nonpositive, no splitting will be done.

Splitting slightly increases the memory footprint of the set, but the splits
share storage with the full symbol so the cost is only slices, map entries, and
a bit of processing time. In the encoded form, there will also be additional
entries in the symbol table.

N.B.: Decoders do not need to change; apart from the extra entries in the
symbol table, this change does not affect the semantics.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2180