Add a definition of a durable linking service based on xrefs.
Summary:
Adds link.proto, which defines a LinkService to resolve qualified names to
definition locations in source text. The service is designed to work as a
wrapper on the XRefService, though it could also be implemented separately.

Unlike IdentifierService, the LinkService permits the client to specify
constraints on the qualified name lookup and filters on the results, since
qualified names will in general not be uniquely-defined.

A basic implementation of the LinkService is included.

Reviewers: #core_team, zarko

Reviewed By: #core_team, zarko

Subscribers: jrtom, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2185