KLS: Include documentation text in hover replies.
Summary:
If the documentation reply for a ticket contains comment text, include it in
the reply to a request for hoverdocs. Linkage markup is stripped from the text
if it is present, as LSP affords us no way to encode it in the result.

Bump max allowed Bazel version from 0.10.0 to 0.10.1.

Reviewers: salguarnieri, #core_team

Reviewed By: salguarnieri, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2187