go: Generate separate package init functions per file.
Summary:
Previously, we were blaming all package-level function calls on an implicit
init function for which the indexer created one instance per package on demand.
But packages may consist of multiple files, so calls may wind up blamed on a
function which appears to be located in a different file. This isn't strictly
wrong, but it is confusing to the user.

To address this, allow there to be a separate implicit package init function
per package source file, as needed. These are disambiguated by the offset in
the token stream of the package clause in that file, which is unique within the
package. Calls within a given file are mapped to that file's package init
instead of a shared one, ensuring that each gets attributed to the correct
file.

Reviewers: schroederc, #core_team

Reviewed By: schroederc, #core_team

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2197