Handle qualified names at the root of MarkedSource.
Summary:
Prior to this commit, a MarkedSource message consisting of only an identifier
node at the root of the tree would produce an empty result, because it has no
child nodes to recur into. Rework the extraction to use the root as a base case.

Along the way, drop the overly-complex stream-based child traversal in favour
of a simple loop.

Add a test for a root identifier.

Reviewers: craigbarber, #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2203