KLS: Fall back to document open on non-matching changes.
Summary:
When the LSP client posts a change for a document the server does not have
open, fall back to opening the file instead. Since the server uses full sync
for all changes anyway (to handle patching), this works around the possible
case where the server may exit and be restarted by the environment, losing its
knowledge of previously-opened files.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2228