KLS: Patch target references for files known to the server.
Summary:
Ordinarily it is not possible to patch xref targets, since doing so would
require fetching the target files. But for the special case where the target
file is already open in the KLS (e.g., it's for a file that is being edited in
the current workspace), it's safe to rewrite target references based on the
patch data from those targets.

This change doesn't cause any files to be opened that weren't already.
Based on an implementation proposed by Amol Mandhane.

Reviewers: #core_team, jrtom

Reviewed By: #core_team, jrtom

Subscribers: jrtom, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2230