Give Kythe URI libraries proto visibility.
Summary:
Although these libraries are not proto libraries themselves, it is not
practical to make much use of Kythe protos without Kythe URI parsing.  Thus:
Give these the same access visibility as the protos so that a client can depend
on these libraries without having to depend on everything else.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2235