Fix issues with $D in newer versions of Bazel and include path subsetting in clang 4.0.1
Summary:
Fix issues with $D in newer versions of Bazel and include path subsetting in clang 4.0.1

Bazel 0.11.1 no longer likes just $D as "common build directory for all outputs" and the replacement is "$(dirname out)" if there is a single output or always the package root if there is more than one.  This is more predictable than the previous behavior, but not backwards compatible.

Clang 4.0.1 outputs a different list of include path for `-xc` and `-xc++` but Bazel uses the same list for both modes (and switches modes by file extension), leading to breakage when compiling plain-C which uses system headers.

Test Plan:
It would be appreciated if someone on an older Bazel can test this with
bazel build @net_zlib//:all

Reviewers: ronshapiro, fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2238