First attempt at implementing a generalized extraction configuration schema and parsing logic.
Summary:
(This is an incremental CL)
The goal here was to implement an intermediate configuration schema which can be used to generate a customized extraction container tailored for the individual repo. This schema would be the output of both the automated repo examination logic as well as a parser for a separate user friendly schema which could live in the repo itself. In both cases, the common patterns you see in the mvn_config.json file could be stored in a mapping used to generate this intermediate form.

Reviewers: danielmoy-google, #core_team, fromberger

Reviewed By: danielmoy-google, #core_team, fromberger

Subscribers: fromberger, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2243