serving/.../explore.go: fix Parents/Children impls to properly skip input tickets with no table entries
Summary:
The original intent was to skip input tickets for which there was no data.
Because we weren't, the Children API was failing because it was reading a default 'type' value of PARENTS out of the empty Relatives response.
(We may want to consider creating an INVALID/UNKNOWN entry for the Relatives.type field, to make diagnosis easier.)

Reviewers: fromberger

Reviewed By: fromberger

Subscribers: #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2265