First pass at a smoke test harness for repo extraction.
Summary:
There's open issues in this commit which should be addressed in future
commits:

1. currently relies on freebie quota, instead this needs to fix x_oauth2
import and call with a proper github token.

2. directory paths don't match up. needs some improvement to do more
fuzzy matching (be tolerant to the first few path elements not matching
up)

3. reading from github shouldn't be blocked by failed extraction -
should consider doing in parallel or something.  Alternatively, leave
this as-is because github quota exhaustion is a problem - might as well
not read from github individually if the clone fails

4. split apart, not have all logic in one giant file

Reviewers: fromberger, schroederc, #core_team, craigbarber

Reviewed By: #core_team, craigbarber

Subscribers: schroederc, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2293