Specify and implement the new "kzip" storage format.
Summary:
This change documents and implements the proposed kzip format, which is
intended to eventually replace .kindex files and index packs as a storage
medium for Kythe compilation records. The kzip format is similar to an index
pack, but instead of compressing each file individually the whole directory
structure is packed into a ZIP archive.

There are several related changes here:

- Add IndexedCompilation to analysis.proto.
- Update generated proto messages.
- Add documentation for the kzip format (asciidoc, spec).
- Allow go_test to support "external" tests.
- Add a Go implementation of the kzip format.
- Implement unit tests for the Go kzip package.

Reviewers: #core_team, shahms

Reviewed By: #core_team, shahms

Subscribers: shahms, #core_team

Differential Revision: https://phabricator-dot-kythe-repo.appspot.com/D2324