Use nicer permissions for kzip file entries.
This is a cosmetic change to make the Go kzip library use more friendly
permission bits for the files it stores into a kzip archive. This has no effect
on library use, but makes things a little nicer to unpack manually for
debugging purposes.

The root directory is now defaulted to 0755, and content files to 0600.