Make the schema package support "go get" again. (#55)
* Make the schema package support "go get" again.

Addresses #54.

Prior to this change, the schema package has some logic in schema.go that
depends on a Bazel-generated file. This prevents the package from being
go-gotten unless we check in that file. So we'll check a copy in.

To do this, update the existing genrule, which worked by concatenating the text
proto directly into a source file, with a tool that generates Go source that
doesn't require the text proto. The file emitted by the new tool is independent
of the rest of the package, and includes both the constant maps (read from the
text proto at compile time) and the functions that use them.  This also makes
it easier to update the file without invoking Bazel.

This change also adds a rule that will fail to build if the generated code used
by the Bazel rule for the schema library differs from the checked-in version of
the data file provided for users of "go get".

It also includes a script to bring things back into sync.

* Convert the sync-check rule into a sh_test.

Instead of making the rule fail to build when the files differ, make a trivial
test fail. This is mostly cosmetic, but tests give a lot nicer output when they
fail than builds do.

* Add the indexdata.go file to the repo. Oops.

* Push file-diff test logic into a macro.

We can use this to check for differences in the generated proto files that we
ship (but I'll do that in a future change).

* Fix copyright notice issues.

- Don't vary the date in the generated output.
- Add the missing notice to the generator.