Bazel extractor: Add support for kzip output.
Add a new ExtractToFile method that writes its result to a kzip instead of
returning a kindex compilation. The existing Extract method is unchanged in its
effect.  In support of this, rework the fixup mechanism to take a compilation
unit instead of a kindex compilation. For now the existing fixup continues to
work until I can migrate the existing usage; once that's done I'll remove the
old one.

The Extract and ExtractToFile methods share most of their common plumbing for
processing the action info and reading file inputs.

Other related changes here:

 - Unexport the classifyInputs method.
 - Unexport the fetchInputs method.
 - Pass a proto rather than a compilation to classifyInputs.
 - Remove dependency on the kindex package's AddDetails method.
 - Remove unused build dependencies.
 - Update unit tests for ExtractToFile.

The main changes to the tests are to pull apart the common pieces of the
original Extract test and use them to test both Extract and ExtractToFile. The
functional differences are minor, mainly this just hoists the test data to the
package level and to drop value checking into helpers.