Support imputes and inline metadata. (#70)
Support imputes and inline metadata.

This PR adds a new metadata rule, anchor_anchor,
which makes the `imputes` edge useful. It also allows
for metadata to be included in the same file that it
describes using a new pragma. (A test is included that
documents a second way of doing this using __FILE__ that
would be difficult to do in practice, owing to having to
get the offsets right).