kzip: Allow propagating Close. (#130)
Add a new constructor NewWriteCloser, which behaves as the normal constructor
save that it takes an io.WriteCloser instead of an io.Writer, and arranges that
when closing the kzip writer, the underlying io.WriteCloser will also be closed.

This simplifies use of the library with files that must otherwise be closed
explicitly. Tests are included to verify that error propagation does the
sensible thing.