properly handle utf16 surrogate pairs (#2815)
The 'utf8' module maps nodejs UTF-16 offsets <-> Kythe UTF-8 offsets.
Because of surrogate pairs, a single code point in UTF-8 may be multiple
UTF-16 code units.

The fix also includes a unit test for the module that includes some
surrogate pairs.  To run this unit test we bring in a dependency on
Jasmine and adjust the test-running machinery to run both the unit
test and the end-to-end test.

This code was mostly written by Bowen Ni <bowen1897@gmail.com>; I just
added the Jasmine glue for running as part of open source Kythe.