Go: Fix declarations of anonymous struct type fields. (#2828)
Addresses #2827. Previously, the code to do this depended on there being a
composite literal to initialize the value. But that isn't always present.  When
it isn't, the expression type itself will have the information we need.

The interesting cases are:

    var v struct { X int }

and

    var w = struct { X int }{X: 25}

Both are possible, but in that case we want the first case to dominate.

I updated the existing tests to cover this better, and added a test to catch
the specific case that motivated #2827.