Add ProcessAnnotation later.
The process for gathering annotations was:
1. Add ProcessAnnotation
2. Check for anything at --processors
3. If there's nothing then check for processors exposed by META-INF

However (3) will never happen because in (1) we always add a processor. Therefore if we move (1) to happen later we have a chance of getting to annotations.

This is necessary because some processors generate source code and Kythe wasn't running those processors resulting in broken code.