RDF: Encode /kythe/code facts in Base64.
Fixes #2855. Per https://www.w3.org/TR/2014/REC-n-triples-20140225, the syntax
of N-Triples is expressed over UTF-8 encoded Unicode code points.

We are escaping code sequences correctly, but not binary values.  The
/kythe/code fact carries a wire-format protobuf message, which is not UTF-8 and
must therefore be escaped. This commit checks specifically for that fact when
encoding triples, and Base64 encodes the value. This matches the JSON encoding
for proto3 messages.