bazel_go_extractor: Update for current bazel_go rules.
This program was not updated since we switched away from the legacy Go rules,
and was still expecting the build to generate a symlink forest and other such
nonsense that no longer applies.

Replace this with a much simpler implementation. The action's command line now
includes the file locations directly, so we do not need to guess and remap path
prefixes. Most of the old code was simply deleted.

Fixes #2912.