KLS: Deal more gracefully with zero-width definition spans. (#2944)
If an indexer emits a zero-width span by providing a start location without an
end, the language server will currently discard it because it lacks an end
marker. While this is not an ideal way for the indexer to represent
definitions, we can get a somewhat less surprising result by reusing the start
location as the end when the latter wasn't provided.