Force GoDetails to be loaded in kzip c++ tests
kzip_reader_test and kzip_writer_test were failing on macOS because they
couldn't load the GoDetails message. I think the message needs to appear
somewhere in the code for the proto loading code to add it to the list of
available types. So I added a dummy instance of GoDetails to every test that
tries to read it. I think we actually only need one per file, but one per test
is more explicit and keeps the hack closer to the reason it exists.