Use our own custom shasum tool (#2983)
shasum is a perl script, which means it is tricky to put in the bazel
sandbox. On Linux, you execute shasum, but on macOS apparently the perl
wrapper finds the right version of the shasum and executes that. Since
we already have our own shasum binary, we should just use that and make
this problem go away.

Part of #2967