Default extracted verifier tests to not check for singletons.
The examples embedded in the schema and other related docs have a lot of
variables that are used only once (i.e., singletons), which trip on the
verifier now that the default is to flag singletons.

For example purposes it's fine to have singletons -- the check is there to
avert typographical errors in unit tests. Errors could also happen in the
examples, but the unit tests should be catching any real bugs that might be
disguised in that case.