Remove further usage of deprecated repository rules
These rules are being deprecated in favour of the starlark variations.
This fixes the remaining rules that were missed in a previous PR.

My change in #2952 was incomplete and not tested correctly, I hadn't
correctly setup my environment and was assuming the unrelated failure
meant that the original repository loading was passing. The evaluation
of the rules is lazy and doesn't immediately trigger the incompatible
change warning.

Tested to build completion with `bazel build
//... --incompatible_remove_native_git_repository
--incompatible_remove_native_http_archive`.