Web: Unpack zip files when staging documentation.
Now that the asciidoc rules emit .zip files, we need to unpack those when
staging the web site for deployment. After rsync, unpack the .zip archives and
then remove them so they don't get pushed out to the site. It would be harmless
to deploy them, but a waste of space.