Make the asciidoc rule emit outputs to a zip file. (#3013)
Because asciidoc may generate an unknown number of output files (as the result of running filters),
fix the rule to capture these in a ZIP file instead. Also add some documentation so that future readers
can see what's going on. This addresses an issue where recent versions of Bazel, which have become
more aggressive about cleaning up the sandbox after an action executes, would lose the support files
from the asciidoc filters.

Also update the website documentation builder to account for the above.