Always build gradle clean. (#3022)
Note this doesn't affect most normal operation since we would tend to
run this inside of a docker container that sets up a clean environment
every time.  However, for both local testing and possibly future usage,
we might not want to assume that - so do a 'gradle clean build' instead
of 'gradle build'.  This means if we're re-using the environment for
whatever reason (maybe to make it cheaper and less overhead? who knows),
we won't sometimes mysteriously get no extraction results.