Bazel extractor: Remove a redundant call to config.FixUnit.
The underlying extraction machinery already invokes the fixup, so it is not
necessary to repeat it when extracting to a file. It is harmess apart from
being wasteful of resources to do so as long as the fixup is idempotent -- but
since that is not required, so don't (unintentionally) rely on it.