Bazel extractor: Avoid re-recording input files already added.
When processing input files, filters may re-introduce files that were already
examined. Prior to this commit, this has two downsides: One, it's redundant
(which, though harmless, is wasteful), and two it can result in entries that do
not get correctly updated (which is incorrect).

Verify that we only add a new required_input for new file paths, and update the
tests to ensure this gets checked.