Don't crash if we try to write the same file to a kzip more than once (#3100)
It looks like we may see the same file multiple times during extraction. If we try to add the file to the zip again, the zip library will throw a generic zip exception. We can avoid this situation by just not appending a file if we know we have already added it to the zip.