Java indexer: do not include Symbol modifiers in hashes (#3139)
Symbol modifiers can be inconsistent between definition/referencing compilations.  The existing exception was found due to the way Turbine handles the 'final' modifier on enums, but it can also be case that other modifiers (like 'strictfp') are inconsistent.  This change makes the symbols hashing less precise but shouldn't affect any known use-case for the indexer.