Add check for asciidoctor gem
If you don't have the asciidoctor gem installed, you can generate a website,
but it will be missing toc/title/etc. because the commands from the script that
use the asciidoctor gem to find those values will silently fail.

Also bump the jekyll and asciidoctor versions to their latest in the gemfile