fix: recover file VNames using file content; order singleton checking correctly
The verifier was initially used to process a single file's subgraph.
As such, it only used begin and end offsets to represent anchor ranges.
Now that it supports multiple files, it ought to disambiguate beween
ranges in different files. Unfortunately, if the verifier is used in
the normal manner (without --use_file_nodes), there is no direct way to
recover all of the VName fields necessary to build anchor goals, since
the filename passed to the verifier may have no relation to what an
indexer produces. This change works around this problem by searching for
file nodes with the same *content* as the verifier input files and using
their VNames to construct anchor goals.

The problem is exacerbated if you are a heavy user of goal groups,
as backtracking might otherwise undo the wrong anchor being selected
for a variable.

BREAKING CHANGE: Particularly if your tests rely on files that your
indexer does not emit as file nodes (e.g., you supply the verifier
with an out-of-band text file with goals), this change will cause your
tests to fail. For behavior that's similar to what the verifier was
doing before, assert the --nofile_vnames flag.

BREAKING CHANGE: If you were using --use_file_nodes and the (default)
--check_for_singletons, you may discover new singletons in your code,
as previous to this change the singleton check was performed *before*
file nodes were read from the input graph.

Fixes #3157.