feat: Support adding Clang USRs to the graph
Clang tools use USRs ("Unified Symbol Resolution") as ticket-equivalents
(but they're insufficiently specific for Kythe). This PR adds a new flag,
--experimental_usr_byte_size, which (if given a nonzero value) instructs
the indexer to take that many bytes off the SHA1 hash of USRs and
add them as clang/usr nodes with /clang/usr edges pointing at the
C++ nodes to which they refer:

    //- @foo defines/binding Foo
    //- FooUsr /clang/usr Foo
    //- FooUsr=vname(_,"","","","usr").node/kind clang/usr
    void foo() { }

Similar to the way we deal with JVM names, the corpus, path,
and root fields of a usr vname are cleared. Clients are permitted
to write their own USR tickets. The USR value itself is encoded
in capital hex (to match Clang's own internal USR stringification,
modulo the configurable size of the SHA1 prefix).

USRs are added only for NamedDecls that:
  * are FunctionDecls
  * are not under implicit template instantiations
  * are not in a DeclContext inside a function body
  * can actually be assigned USRs from Clang

Future PRs may allow various other NamedDecl subtypes to be granted
USRs, but it is unlikely that we will support assigning USRs to
decls that are not at TU/namespace scope.