refactor(cxx): create kythe/cxx/common:path_utils library that doesn't depend on llvm (#3282)
These utils are needed for the cxx indexer and extractor and will be needed by the proto indexer that's currently being open sourced. Hopefully these filepath utils will one day end up as part of abseil, but in the meantime, a common library in the kythe project is useful to avoid duplicating code.

The CleanPath() and JoinPath() functions are moved from an existing path_utils library in the cxx_extractor directory. The rest of that library might be able to move to the common library as well, but they're not needed yet and currently depend on llvm. This existing library has moved from the `kythe` namespace to `kythe::cxx_extractor`.