refactor: remove deprecated kindex from cxx extractor
KZip has replaced the index pack and kindex formats.

Previously the extractor only wrote kzip files if KYTHE_OUTPUT_FILE was
specified in the environment and ended with '.kzip'. This remains
unchanged, but now the extractor also writes kzip files if
KYTHE_OUTPUT_DIRECTORY is specified - this used to result in .kindex
files being written instead. When given a directory, the extractor names
the .kzip files based on a hash of the compilation unit just as it did
with kindex.

Extractor tests have been updated to use kzip instead of kindex.