feat(objc_indexer): support marked source for @property
Add basic support for generating marked source for properties in
Objective-C. This should offer similar support to what we do for class
members in C++. It doesn't do anything special for the automatically
generated getter and setter methods. It also doesn't do anything special
with the @property string (though it is present in the marked source).