fix: change tools/modules/update.sh to use an archive instead of clone
One way of fixing #3307.

This will make the initial download of llvm and clang faster, by not
requiring a clone of the entire ~3GB repo.  It also save on the ongoing
disk used.  However, it does make updates marginally slower because you
have to re-download a new archive.  Given that updates are vastly
dominated by the cost of building llvm&clang, this is probably ok.

Also this has the side effect of making images for Google Cloud Build +
bazel significantly smaller, which will be nice for #3210.