feat: add a simple kzip validator
Opens a local kzip, and compares it to the contents in a local directory
repo root.  Prints out how much is missing, and isolates the top missing
directories and prints them as well.

Usage
  kzip_validator -kzip <kzip-file> \
                 -local_repo <repo-root-path> \
                 -lang <comma-separated-langs> \
                 [-missing_file <file-to-save-missing-paths]

Future versions of this should support grabbing the repo down from a
remote url instead of loading it from local disk.  Future versions of
this might also want to work off of a blacklist for sub-parts of repos
the user might not care about.