feat: add example for extracting multiple mavens out of guava in gcb (#3332)
* feat: add example for extracting multiple mavens out of guava in gcb

This extracts multiple parts of guava by executing the mvn clean compile
with attached kythe extraction from the top level directory, instead of
just from inside guava/guava like the previous version of this did.

I verified for a local case that it outputs the same hash kzip so
guava/guava the main version should still be there.

Once #3331 is done it'll be easier to verify behavior of this and other
similar examples.  Once  #3328 is done we'll have better handling of the
other parts of guava that are probably still not extracting.