fix(proto_indexer): skip synthesized map entry types (#26)
The proto compiler synthesizes types to represent map entries. For
example, a "map<string, string> my_map" field would cause a type
"MyMapEntry" to be generated. Because it doesn't actually exist in the
source .proto file, we can't create an anchor node for it. Previously we
attempted to anyways and printed errors of the form: "Unexpected
location vector [] while walking..." upon failure. This commit changes
to just skip over them.