feat: add standalone extractor for textproto (#35)
The textproto extractor looks for # proto-file: and # proto-message: comments OR flags to determine what the schema of the textproto file is. The proto extractor then does most of the heavy lifting of tracking down and bundling up all of the proto dependencies.

The tests are using the same kzip diffing script/bzl as the proto extractor tests.