build: update to tagged versions of stringset and shell
Rather than pinning to specific random commit hashes, point to the release tags
for these libraries. This required a small change to the shim, which did not
handle the tag/commit combination quite correctly.

More libraries could probably productively be pinned this way, and it is
generally safer to do so since that's how Go modules will do it too. For now,
though, I'm just doing the ones I know the versions for.