use the first token in the file as the module link (#3260)
When a file imports another file, with syntax like
  import * as x from 'some/path';
we want 'some/path' to refer to a VName that just means "the entire
file".  It doesn't refer to any text in particular, so we just mark
the first letter in the file as the anchor for this.  (Note that
specifically, it does not refer to an on-disk path, so we cannot
use ref/file.  See discussion of "Module name" in README.md.)

Previously we used the first 'export' statement in the file, but this
meant that the xref would jump you to a random spot within the file.