build: remove unused dependency on github.com/google/go-github (#3479)
There are no imports of this package anywhere. If we do want to take a
dependency on it in the future, we should do so at one of the release tags.
For now, doing so at any recent tag will pull in extra dependencies, so I
propose to drop it till we need it concretely.