fix: adding support to gcp extraction generation for subdirs
Also supports multiple extraction targets for a single corpus if the
top-level pom.xml file does not pull in all of the subdirectories.

This commit has a somewhat undesireable intermediate state here where we
haven't settled on where CORPUS is specified.  Right now we support both
${_CORPUS} and having it codified in the repo config file.  I am leaning
towards not supporting ${_CORPUS} and having a config generator that
adds it in if necessary, but that's not addressed in this commit.