fix: quiet the error printing in gcp (#3535)
This is in preparation for making .err/.out just print inline.  Not all of our extractors will probably be able to make separate .err/.out files, so we'll unify to just having them all print errors inline.  The error logs can be parsed easily if we prefix extractor errors, so there's a change forthcoming for that.