feat(api): allow a semantic scope to be associated per decoration (#3707)
The semantic scope for each reference corresponds to the target of an anchor node's childofedge. Currently, these are only prescribed for ref/call anchors, but may be expanded to other anchor kinds in the future. Experimental support for childof edges on other anchors is guarded by the Java indexer's --emit_anchor_scopes flag.