fix(cxx_indexer): insist that compilation units have working directories
The C++ (+ friends) indexer should be hermetic modulo claiming when
reading from extracted compilation units. Currently, it will use the
indexing environment's working directory when one is not set in the
compilation unit (or if the value that was set is not an absolute path).
This PR turns this into an error condition.