Create presubmit.yml
This (+ a PR in https://github.com/bazelbuild/continuous-integration) will put
make Bazel changes that break Kythe visible to the Bazel team before a
release is cut.

This just builds for Ubuntu 18.04, but can be expanded to do windows,
osx, or ubuntu 16.04. For now I want to start small and see if we can get
the two to just talk to each other.