typescript: document separate compilation
This describes how separate compilation works for Google's TS builds.
It's not obvious from this code that this is how things work; for
example, in non-Google TS builds, the "compilation unit" that TS works
with is generally all source code in the app all at once.

(As you might expect, that is costly; recent versions of TypeScript have
been working on adding some more incremental compilation support, but
I'm not clear on how exactly it works.)