fix(typescript_indexer): differentiate signatures of static and instance members (#3819)
Currently class instance and static properties generate the same type of signature, which collide when there is a static and instance member of the same name. This resolves that by scoping the signature of instance members to the class type rather than value.

old:

```
class C {
  //- @prop defines/binding VName("C.prop", _, _, _, _)
  prop;
  // Oh no!
  //- @prop defines/binding VName("C.prop", _, _, _, _)
  static prop;
```

new:

```
class C {
  //- @prop defines/binding VName("C#type.prop", _, _, _, _)
  prop;
  //- @prop defines/binding VName("C.prop", _, _, _, _)
  static prop;
```

Part of compliance for #3779, #3783.